requirements.txt
python-telegram-bot==12.4.2
pymongo==4.0.1
